task-17-re-export: created modules for all classes inside new folder 'classes' + added re-export of classes in index.ts file
