+++ src/classes/index.ts
@@ -1,0 +1,3 @@
+export { default as Encyclopedia } from './encyclopedia';
+export { default as ReferenceItem } from './referenceItem';
+export { default as UniversityLibrarian } from './universityLibrarian';
