+++ src/classes/universityLibrarian.ts
@@ -1,0 +1,13 @@
+import * as Interfaces from '../interfaces';
+
+class UniversityLibrarian implements Interfaces.Librarian {
+    name: string;
+    email: string;
+    department: string;
+
+    assistCustomer(custName: string): void {
+        console.log(`${this.name} is assisting ${custName}`);
+    }
+}
+
+export default UniversityLibrarian;
